refactor hooks to run only for the necessary platforms
