+++ plugin.xml
@@ -11,6 +11,10 @@ xmlns:android="http://schemas.android.com/apk/res/android">
 	</engines>
 
 	<platform name="android">
+		<hook type="after_prepare" src="scripts/android/after_prepare.js" />
+		<hook type="after_plugin_install" src="scripts/android/after_plugin_install.js" />
+		<hook type="before_plugin_uninstall" src="scripts/android/before_plugin_uninstall.js" />
+
 		<js-module name="FirebasePlugin" src="www/firebase.js">
 			<clobbers target="FirebasePlugin" />
 		</js-module>
@@ -57,10 +61,13 @@ xmlns:android="http://schemas.android.com/apk/res/android">
 		<framework src="com.google.firebase:firebase-messaging:+" />
 		<framework src="com.google.firebase:firebase-config:+" />
 		<framework src="com.google.firebase:firebase-perf:+" />
-
 	</platform>
 
 	<platform name="ios">
+		<hook type="after_prepare" src="scripts/ios/after_prepare.js" />
+		<hook type="after_plugin_install" src="scripts/ios/after_plugin_install.js" />
+		<hook type="before_plugin_uninstall" src="scripts/ios/before_plugin_uninstall.js" />
+
 		<js-module name="FirebasePlugin" src="www/firebase.js">
 			<clobbers target="FirebasePlugin" />
 		</js-module>
@@ -111,8 +118,4 @@ xmlns:android="http://schemas.android.com/apk/res/android">
 			<clobbers target="FirebasePlugin" />
 		</js-module>
 	</platform>
-
-	<hook src="scripts/after_prepare.js" type="after_prepare" />
-	<hook src="scripts/after_plugin_install.js" type="after_plugin_install" />
-	<hook src="scripts/before_plugin_uninstall.js" type="before_plugin_uninstall" />
 </plugin>
