--- scripts/after_prepare.js
@@ -1,119 +1,0 @@
-#!/usr/bin/env node
-
-'use strict';
-
-/**
- * This hook makes sure projects using [cordova-plugin-firebase](https://github.com/arnesson/cordova-plugin-firebase)
- * will build properly and have the required key files copied to the proper destinations when the app is build on Ionic Cloud using the package command.
- * Credits: https://github.com/arnesson.
- */
-var fs = require('fs');
-var path = require('path');
-
-fs.ensureDirSync = function (dir) {
-  if (!fs.existsSync(dir)) {
-    dir.split(path.sep).reduce(function (currentPath, folder) {
-      currentPath += folder + path.sep;
-      if (!fs.existsSync(currentPath)) {
-        fs.mkdirSync(currentPath);
-      }
-      return currentPath;
-    }, '');
-  }
-};
-
-var config = fs.readFileSync('config.xml').toString();
-var name = getValue(config, 'name');
-
-var IOS_DIR = 'platforms/ios';
-var ANDROID_DIR = 'platforms/android';
-
-var PLATFORM = {
-  IOS: {
-    dest: [
-      IOS_DIR + '/' + name + '/Resources/GoogleService-Info.plist',
-      IOS_DIR + '/' + name + '/Resources/Resources/GoogleService-Info.plist'
-    ],
-    src: [
-      'GoogleService-Info.plist',
-      IOS_DIR + '/www/GoogleService-Info.plist',
-      'www/GoogleService-Info.plist'
-    ]
-  },
-  ANDROID: {
-    dest: [
-      ANDROID_DIR + '/google-services.json',
-      ANDROID_DIR + '/app/google-services.json'
-    ],
-    src: [
-      'google-services.json',
-      ANDROID_DIR + '/assets/www/google-services.json',
-      'www/google-services.json',
-      ANDROID_DIR + '/app/src/main/google-services.json'
-    ],
-  }
-};
-
-function copyKey (platform) {
-  for (var i = 0; i < platform.src.length; i++) {
-    var file = platform.src[i];
-    if (fileExists(file)) {
-      try {
-        var contents = fs.readFileSync(file).toString();
-
-        try {
-          platform.dest.forEach(function (destinationPath) {
-            var folder = destinationPath.substring(0, destinationPath.lastIndexOf('/'));
-            fs.ensureDirSync(folder);
-            fs.writeFileSync(destinationPath, contents);
-          });
-        } catch (e) {
-          // skip
-        }
-      } catch (err) {
-        console.log(err);
-      }
-
-      break;
-    }
-  }
-}
-
-function getValue(config, name) {
-  var value = config.match(new RegExp('<' + name + '(.*?)>(.*?)</' + name + '>', 'i'));
-  if (value && value[2]) {
-    return value[2]
-  } else {
-    return null
-  }
-}
-
-function fileExists (path) {
-  try {
-    return fs.statSync(path).isFile();
-  } catch (e) {
-    return false;
-  }
-}
-
-function directoryExists (path) {
-  try {
-    return fs.statSync(path).isDirectory();
-  } catch (e) {
-    return false;
-  }
-}
-
-module.exports = function (context) {
-  //get platform from the context supplied by cordova
-  var platforms = context.opts.platforms;
-  // Copy key files to their platform specific folders
-  if (platforms.indexOf('ios') !== -1 && directoryExists(IOS_DIR)) {
-    console.log('Preparing Firebase on iOS');
-    copyKey(PLATFORM.IOS);
-  }
-  if (platforms.indexOf('android') !== -1 && directoryExists(ANDROID_DIR)) {
-    console.log('Preparing Firebase on Android');
-    copyKey(PLATFORM.ANDROID);
-  }
-};
