+++ scripts/lib/utilities.js
@@ -1,8 +1,20 @@
 /**
  * Utilities and shared functionality for the build hooks.
  */
-
+var fs = require('fs');
 var path = require("path");
+
+fs.ensureDirSync = function (dir) {
+  if (!fs.existsSync(dir)) {
+    dir.split(path.sep).reduce(function (currentPath, folder) {
+      currentPath += folder + path.sep;
+      if (!fs.existsSync(currentPath)) {
+        fs.mkdirSync(currentPath);
+      }
+      return currentPath;
+    }, '');
+  }
+};
 
 module.exports = {
 
@@ -38,4 +50,53 @@ module.exports = {
     return path.join("platforms", "ios", appName + ".xcodeproj", "project.pbxproj");
   },
 
+  copyKey: function (platform) {
+    for (var i = 0; i < platform.src.length; i++) {
+      var file = platform.src[i];
+      if (this.fileExists(file)) {
+        try {
+          var contents = fs.readFileSync(file).toString();
+
+          try {
+            platform.dest.forEach(function (destinationPath) {
+              var folder = destinationPath.substring(0, destinationPath.lastIndexOf('/'));
+              fs.ensureDirSync(folder);
+              fs.writeFileSync(destinationPath, contents);
+            });
+          } catch (e) {
+            // skip
+          }
+        } catch (err) {
+          console.log(err);
+        }
+
+        break;
+      }
+    }
+  },
+
+  getValue: function (config, name) {
+    var value = config.match(new RegExp('<' + name + '(.*?)>(.*?)</' + name + '>', 'i'));
+    if (value && value[2]) {
+      return value[2]
+    } else {
+      return null
+    }
+  },
+
+  fileExists: function (path) {
+    try {
+      return fs.statSync(path).isFile();
+    } catch (e) {
+      return false;
+    }
+  },
+
+  directoryExists: function (path) {
+    try {
+      return fs.statSync(path).isDirectory();
+    } catch (e) {
+      return false;
+    }
+  }
 };
